update commit
Résultats des sessions dans directory results.
vroum.c to logic.c
Message d'aide update
Ajustement du nombre de tours pour chaque sessions.
Mise à jour des num des pilotes pour cette année
