--- Formula_one-main/P1.csv
@@ -1,24 +1,0 @@
-P1
-id; best s1; best s2 ; best s3 ; best lap; total time
-| id |   s1  |   s2  |   s3  | best lap | best s1 | best s2 | best s3 | pitstop | crash 
-|----|-------|-------|-------|----------|---------|---------|---------|---------|-------
-| 10 | 28.31 | 37.23 | 41.91 |   79.97 |  27.15 |  25.91 |  25.38 | 3 |false |
-| 24 | 29.94 | 34.98 | 29.65 |   81.34 |  26.84 |  25.43 |  25.53 | 2 |false |
-| 16 | 29.01 | 27.70 | 25.23 |   81.85 |  26.55 |  25.01 |  25.16 | 2 |false |
-| 55 | 43.55 | 35.71 | 36.78 |   83.81 |  25.22 |  25.02 |  26.25 | 3 |false |
-| 63 | 44.48 | 43.00 | 41.20 |   84.79 |  26.87 |  25.37 |  26.72 | 1 |false |
-| 18 | 29.24 | 44.52 | 26.34 |   84.82 |  28.92 |  25.34 |  25.83 | 2 |false |
-| 77 | 26.11 | 40.75 | 28.94 |   85.61 |  25.56 |  28.97 |  26.94 | 2 |false |
-| 31 | 40.76 | 33.86 | 43.52 |   85.82 |  28.65 |  25.30 |  25.14 | 1 |false |
-|  1 | 37.27 | 41.54 | 26.94 |   87.01 |  25.65 |  28.25 |  25.30 | 2 |false |
-| 44 | 44.80 | 42.95 | 30.48 |   87.17 |  29.84 |  27.27 |  30.06 | 0 |true |
-|  6 | 27.85 | 43.59 | 39.70 |   88.65 |  27.85 |  26.45 |  25.96 | 0 |false |
-| 22 | 26.69 | 37.25 | 25.14 |   89.08 |  26.69 |  37.25 |  25.14 | 0 |true |
-| 11 | 39.65 | 37.93 | 32.84 |   91.22 |  25.10 |  27.46 |  26.23 | 2 |true |
-|  3 | 25.88 | 35.39 | 30.99 |   91.69 |  25.88 |  27.15 |  25.01 | 0 |true |
-|  4 | 27.62 | 26.77 | 38.60 |   91.81 |  26.50 |  26.63 |  25.04 | 3 |false |
-|  9 | 38.20 | 28.82 | 31.36 |   91.86 |  29.21 |  26.34 |  26.27 | 1 |false |
-| 14 | 28.78 | 30.88 | 44.13 |   91.97 |  26.66 |  25.05 |  26.51 | 2 |false |
-| 47 | 31.11 | 28.23 | 40.56 |   92.79 |  25.82 |  28.23 |  26.29 | 0 |true |
-|  5 | 37.73 | 35.18 | 29.15 |   98.47 |  25.17 |  27.42 |  25.58 | 1 |false |
-| 23 | 44.10 | 31.34 | 39.91 |  115.35 |  44.10 |  31.34 |  39.91 | 0 |true |
