--- Formula_one-main/P2.csv
@@ -1,24 +1,0 @@
-P2
-id; best s1; best s2 ; best s3 ; best lap; total time
-| id |   s1  |   s2  |   s3  | best lap | best s1 | best s2 | best s3 | pitstop | crash 
-|----|-------|-------|-------|----------|---------|---------|---------|---------|-------
-|  4 | 40.52 | 37.04 | 42.41 |   83.04 |  25.74 |  26.83 |  26.93 | 2 |false |
-| 63 | 37.39 | 33.27 | 25.02 |   85.56 |  26.86 |  26.16 |  25.02 | 3 |false |
-|  6 | 40.76 | 33.86 | 43.52 |   85.82 |  28.65 |  25.30 |  25.14 | 1 |false |
-|  9 | 31.11 | 39.09 | 35.18 |   86.58 |  25.05 |  25.34 |  25.65 | 4 |false |
-|  5 | 30.64 | 25.45 | 32.97 |   89.06 |  26.02 |  25.45 |  25.23 | 2 |false |
-| 11 | 31.34 | 35.91 | 36.28 |   89.47 |  28.56 |  25.98 |  26.08 | 3 |false |
-| 31 | 33.67 | 44.13 | 27.33 |   89.57 |  25.85 |  25.22 |  27.20 | 1 |false |
-| 10 | 41.22 | 27.50 | 25.73 |   89.58 |  25.92 |  27.50 |  25.73 | 1 |false |
-| 55 | 36.46 | 25.98 | 40.60 |   89.66 |  26.02 |  25.98 |  25.38 | 1 |false |
-|  1 | 30.18 | 31.80 | 30.76 |   89.67 |  26.39 |  25.26 |  25.48 | 3 |false |
-| 14 | 41.69 | 41.14 | 27.94 |   90.48 |  25.17 |  28.37 |  27.94 | 1 |false |
-| 18 | 43.89 | 39.14 | 28.50 |   90.73 |  27.98 |  27.90 |  28.50 | 0 |true |
-| 24 | 42.85 | 25.25 | 33.47 |   91.54 |  26.40 |  25.25 |  28.09 | 1 |false |
-| 22 | 31.11 | 28.23 | 40.56 |   92.79 |  25.82 |  28.23 |  26.29 | 0 |true |
-| 16 | 41.92 | 37.97 | 29.05 |   93.93 |  25.87 |  29.64 |  25.53 | 3 |true |
-| 77 | 29.95 | 30.14 | 36.59 |   94.18 |  28.68 |  26.71 |  26.65 | 0 |true |
-| 44 | 25.18 | 33.46 | 44.52 |   96.05 |  25.18 |  25.16 |  29.50 | 3 |false |
-| 23 | 32.74 | 36.84 | 42.91 |   99.07 |  27.69 |  26.71 |  26.87 | 0 |false |
-|  3 | 36.43 | 33.74 | 30.24 |  100.41 |  31.18 |  33.74 |  30.24 | 0 |true |
-| 47 | 28.53 | 38.51 | 42.86 |  102.25 |  25.75 |  27.01 |  38.62 | 1 |true |
