--- Formula_one-main/P3.csv
@@ -1,24 +1,0 @@
-P3
-id; best s1; best s2 ; best s3 ; best lap; total time
-| id |   s1  |   s2  |   s3  | best lap | best s1 | best s2 | best s3 | pitstop | crash 
-|----|-------|-------|-------|----------|---------|---------|---------|---------|-------
-|  9 | 43.32 | 38.89 | 35.68 |   80.22 |  25.36 |  26.58 |  26.36 | 1 |false |
-| 11 | 43.55 | 35.71 | 36.78 |   83.81 |  25.22 |  25.02 |  26.25 | 3 |false |
-| 44 | 37.39 | 33.27 | 25.02 |   85.56 |  26.86 |  26.16 |  25.02 | 3 |false |
-| 10 | 33.43 | 27.30 | 26.23 |   86.97 |  27.12 |  27.30 |  26.23 | 1 |false |
-| 55 | 27.77 | 32.99 | 40.54 |   87.07 |  27.72 |  25.31 |  26.21 | 3 |true |
-| 16 | 34.13 | 37.77 | 29.54 |   87.59 |  28.50 |  25.17 |  26.26 | 1 |false |
-| 18 | 34.36 | 34.59 | 30.65 |   88.95 |  28.77 |  26.26 |  25.74 | 1 |false |
-| 14 | 33.90 | 40.95 | 28.44 |   89.06 |  25.27 |  25.51 |  25.92 | 0 |false |
-| 22 | 34.63 | 30.28 | 42.85 |   89.64 |  28.99 |  29.75 |  25.96 | 0 |true |
-| 24 | 35.06 | 25.05 | 33.97 |   90.12 |  25.74 |  25.05 |  26.72 | 2 |false |
-| 77 | 31.23 | 30.82 | 33.25 |   90.15 |  29.92 |  25.38 |  27.26 | 1 |false |
-|  6 | 32.97 | 33.66 | 44.02 |   90.63 |  28.21 |  25.67 |  25.08 | 1 |false |
-| 47 | 27.74 | 38.50 | 36.67 |   91.28 |  26.36 |  25.92 |  26.36 | 3 |false |
-| 23 | 30.32 | 35.26 | 35.49 |   91.53 |  26.89 |  25.05 |  29.74 | 0 |true |
-|  5 | 42.85 | 25.25 | 33.47 |   91.54 |  26.40 |  25.25 |  28.09 | 1 |false |
-|  3 | 31.11 | 28.23 | 40.56 |   92.79 |  25.82 |  28.23 |  26.29 | 0 |true |
-| 31 | 25.88 | 43.93 | 27.83 |   93.74 |  25.88 |  26.69 |  26.27 | 1 |false |
-|  1 | 42.39 | 31.61 | 31.26 |   95.83 |  25.08 |  26.14 |  25.00 | 3 |false |
-|  4 | 32.74 | 36.84 | 42.91 |   99.07 |  27.69 |  26.71 |  26.87 | 0 |false |
-| 63 | 27.05 | 35.05 | 40.50 |  102.60 |  27.05 |  35.05 |  40.50 | 0 |true |
