--- Formula_one-main/Q1.csv
@@ -1,24 +1,0 @@
-Q1.csv
-id; best s1; best s2 ; best s3 ; best lap; total time
-| id |   s1  |   s2  |   s3  | best lap | best s1 | best s2 | best s3 | pitstop | crash 
-|----|-------|-------|-------|----------|---------|---------|---------|---------|-------
-| 63 | 30.25 | 30.20 | 41.45 |   79.32 |  25.90 |  26.18 |  25.73 | 2 |false |
-| 10 | 35.58 | 44.70 | 43.60 |   83.37 |  25.82 |  25.09 |  25.36 | 0 |false |
-|  6 | 29.33 | 43.96 | 43.28 |   83.70 |  26.10 |  25.11 |  25.93 | 2 |false |
-| 31 | 40.96 | 25.16 | 37.91 |   83.85 |  25.10 |  25.16 |  25.02 | 3 |false |
-| 22 | 44.24 | 35.55 | 40.80 |   84.45 |  27.32 |  25.58 |  26.43 | 5 |true |
-| 11 | 34.22 | 26.97 | 37.28 |   84.75 |  25.94 |  26.09 |  25.65 | 2 |false |
-| 47 | 41.40 | 25.20 | 41.55 |   85.27 |  27.10 |  25.20 |  25.39 | 2 |false |
-| 55 | 31.70 | 32.57 | 32.92 |   85.82 |  25.92 |  25.50 |  26.07 | 3 |false |
-|  4 | 39.22 | 33.08 | 27.15 |   87.10 |  31.05 |  25.29 |  25.70 | 1 |true |
-|  1 | 40.58 | 36.49 | 30.40 |   87.63 |  26.65 |  25.47 |  26.05 | 1 |false |
-| 44 | 42.24 | 40.07 | 37.36 |   88.84 |  26.36 |  25.66 |  25.81 | 3 |false |
-| 14 | 40.08 | 30.13 | 26.58 |   89.54 |  25.67 |  25.23 |  26.58 | 3 |false |
-| 18 | 43.83 | 25.91 | 28.92 |   89.56 |  25.73 |  25.50 |  26.67 | 1 |false |
-|  9 | 43.63 | 27.09 | 33.67 |   91.00 |  26.38 |  27.09 |  25.31 | 2 |false |
-| 24 | 26.89 | 42.14 | 38.78 |   92.29 |  26.89 |  25.69 |  26.22 | 2 |false |
-| 77 | 29.60 | 28.82 | 41.57 |   93.33 |  25.14 |  27.18 |  25.84 | 1 |false |
-| 23 | 40.28 | 37.62 | 43.63 |   93.66 |  26.08 |  25.43 |  26.31 | 3 |true |
-|  5 | 42.98 | 38.62 | 38.49 |   95.40 |  26.84 |  26.02 |  27.32 | 1 |false |
-|  3 | 36.42 | 37.84 | 30.94 |  105.19 |  36.42 |  37.84 |  30.94 | 0 |true |
-| 16 | 42.89 | 44.19 | 39.96 |  126.79 |  42.83 |  41.03 |  39.96 | 0 |true |
