--- Formula_one-main/Q2.csv
@@ -1,19 +1,0 @@
-Q2.csv
-id; best s1; best s2 ; best s3 ; best lap; total time
-| id |   s1  |   s2  |   s3  | best lap | best s1 | best s2 | best s3 | pitstop | crash 
-|----|-------|-------|-------|----------|---------|---------|---------|---------|-------
-| -1 | 34.39 | 36.36 | 44.80 |   80.44 |  29.55 |  25.53 |  25.24 | 1 |true |
-| 27 | 28.84 | 25.73 | 28.62 |   83.19 |  26.10 |  25.73 |  27.16 | 1 |false |
-| 31 | 44.44 | 39.70 | 34.31 |   85.81 |  28.77 |  25.82 |  26.35 | 1 |false |
-| -1 | 35.62 | 27.37 | 41.01 |   86.21 |  29.03 |  25.31 |  25.08 | 1 |false |
-| 38 | 28.71 | 25.26 | 32.88 |   86.85 |  25.35 |  25.26 |  32.88 | 0 |false |
-| -1 | 35.90 | 42.68 | 33.54 |   87.81 |  26.19 |  25.37 |  26.29 | 2 |false |
-| -1 | 34.03 | 28.55 | 26.75 |   89.32 |  29.63 |  25.71 |  26.75 | 1 |false |
-| -1 | 30.51 | 44.36 | 35.64 |   93.10 |  26.83 |  29.58 |  27.61 | 1 |false |
-| 32 | 34.45 | 35.31 | 28.34 |   93.32 |  25.68 |  26.16 |  28.34 | 0 |false |
-| 28 | 41.60 | 40.14 | 42.48 |   93.44 |  25.79 |  26.28 |  26.13 | 0 |false |
-| 26 | 40.44 | 32.14 | 30.79 |   94.50 |  25.49 |  28.45 |  26.13 | 1 |false |
-| 35 | 33.10 | 26.02 | 38.89 |   94.52 |  25.30 |  25.72 |  26.32 | 0 |false |
-| 44 | 25.80 | 40.10 | 35.48 |   94.67 |  25.80 |  25.83 |  28.89 | 0 |false |
-| 25 | 26.06 | 39.29 | 32.40 |   97.75 |  26.06 |  39.29 |  32.40 | 0 |true |
-| 35 | 44.63 | 27.79 | 39.07 |  111.50 |  44.63 |  27.79 |  39.07 | 0 |true |
