--- Formula_one-main/Q3.csv
@@ -1,14 +1,0 @@
-Q3.csv
-id; best s1; best s2 ; best s3 ; best lap; total time
-| id |   s1  |   s2  |   s3  | best lap | best s1 | best s2 | best s3 | pitstop | crash 
-|----|-------|-------|-------|----------|---------|---------|---------|---------|-------
-| -1094795586 | 31.32 | 39.29 | 33.03 |   80.56 |  27.98 |  25.27 |  27.31 | 0 |false |
-| -1094795586 | 41.33 | 35.95 | 30.22 |   81.92 |  25.50 |  26.38 |  30.04 | 0 |false |
-| -1094795586 | 34.42 | 36.23 | 38.64 |   87.59 |  33.28 |  25.79 |  28.52 | 0 |false |
-| -1094795586 | 25.44 | 35.35 | 34.02 |   88.15 |  25.44 |  26.90 |  25.52 | 0 |false |
-| -1094795586 | 25.41 | 36.49 | 27.68 |   89.58 |  25.41 |  25.51 |  25.38 | 1 |false |
-| -1094795586 | 33.05 | 28.58 | 37.30 |   96.72 |  28.72 |  28.58 |  31.72 | 0 |false |
-| -1094795586 | 26.45 | 38.88 | 39.52 |   97.71 |  26.45 |  31.25 |  25.47 | 1 |false |
-| -1094795586 | 44.06 | 32.08 | 44.60 |  101.17 |  27.42 |  32.08 |  31.51 | 0 |false |
-| -1094795586 | 34.77 | 35.87 | 35.60 |  105.44 |  29.45 |  35.87 |  31.75 | 0 |false |
-| -1094795586 | 43.23 | 40.22 | 42.71 |  111.77 |  34.46 |  32.69 |  38.92 | 0 |false |
