--- Formula_one-main/Race.csv
@@ -1,24 +1,0 @@
-Race.csv
-id; best s1; best s2 ; best s3 ; best lap; total time
-| id |   s1  |   s2  |   s3  | best lap | best s1 | best s2 | best s3 | pitstop | crash 
-|----|-------|-------|-------|----------|---------|---------|---------|---------|-------
-| 10 | 43.51 | 40.62 | 35.28 |   78.56 |  25.26 |  25.23 |  25.04 | 4 |true |
-|  9 | 35.79 | 42.57 | 40.76 |   79.49 |  25.04 |  25.55 |  25.14 | 7 |false |
-| 77 | 39.15 | 32.23 | 34.79 |   80.32 |  25.02 |  25.45 |  25.81 | 3 |true |
-| 44 | 44.72 | 34.62 | 42.00 |   80.34 |  25.00 |  26.04 |  25.11 | 4 |false |
-| 16 | 40.74 | 34.31 | 32.01 |   80.70 |  25.36 |  25.20 |  25.38 | 6 |false |
-|  3 | 30.60 | 35.53 | 27.40 |   81.97 |  26.08 |  25.34 |  25.05 | 4 |true |
-| 55 | 43.82 | 30.54 | 43.28 |   82.96 |  25.65 |  25.73 |  26.08 | 1 |true |
-| 24 | 39.73 | 34.17 | 38.64 |   83.14 |  25.23 |  25.00 |  25.09 | 6 |false |
-| 31 | 34.03 | 25.20 | 37.20 |   83.20 |  25.64 |  25.20 |  25.33 | 4 |true |
-|  1 | 38.59 | 29.59 | 25.17 |   84.33 |  25.97 |  26.06 |  25.17 | 0 |true |
-| 47 | 32.37 | 43.27 | 43.59 |   84.34 |  25.51 |  25.80 |  25.82 | 3 |true |
-|  4 | 37.97 | 35.37 | 28.57 |   85.58 |  25.34 |  25.65 |  25.92 | 1 |true |
-| 11 | 26.63 | 42.58 | 40.97 |   85.71 |  25.27 |  25.13 |  25.07 | 4 |true |
-| 18 | 44.22 | 43.60 | 41.67 |   85.78 |  26.94 |  26.45 |  25.49 | 3 |true |
-| 63 | 42.95 | 42.20 | 29.03 |   86.20 |  25.68 |  25.42 |  25.74 | 7 |false |
-| 23 | 29.18 | 34.92 | 33.19 |   87.15 |  26.67 |  28.00 |  25.87 | 1 |true |
-|  6 | 40.94 | 37.71 | 29.89 |   87.81 |  25.76 |  26.00 |  27.10 | 0 |true |
-| 14 | 37.14 | 29.08 | 35.95 |   88.37 |  25.10 |  25.62 |  28.82 | 1 |true |
-|  5 | 27.88 | 44.83 | 25.34 |   89.96 |  25.11 |  25.20 |  25.34 | 1 |true |
-| 22 | 42.47 | 33.54 | 39.54 |   91.11 |  25.07 |  26.08 |  25.24 | 4 |true |
