--- Formula_one-main/Sprint.csv
@@ -1,24 +1,0 @@
-Sprint
-id; best s1; best s2 ; best s3 ; best lap; total time
-| id |   s1  |   s2  |   s3  | best lap | best s1 | best s2 | best s3 | pitstop | crash 
-|----|-------|-------|-------|----------|---------|---------|---------|---------|-------
-| 22 | 39.34 | 37.74 | 33.29 |   81.57 |  25.24 |  25.97 |  26.69 | 2 |true |
-| 23 | 37.99 | 31.09 | 34.14 |   82.11 |  30.12 |  25.02 |  25.55 | 1 |false |
-| 14 | 30.15 | 30.96 | 26.54 |   83.44 |  29.87 |  28.41 |  25.16 | 1 |false |
-| 16 | 27.45 | 37.65 | 28.23 |   85.07 |  27.45 |  27.92 |  25.25 | 1 |false |
-| 44 | 29.62 | 31.41 | 31.87 |   85.20 |  28.84 |  26.69 |  26.43 | 1 |false |
-| 24 | 36.64 | 44.44 | 34.98 |   85.81 |  30.07 |  26.90 |  26.89 | 0 |false |
-| 10 | 35.55 | 37.56 | 43.16 |   86.67 |  25.41 |  30.68 |  25.27 | 0 |false |
-| 31 | 27.18 | 37.88 | 40.89 |   87.09 |  25.58 |  27.30 |  27.18 | 1 |false |
-| 77 | 25.04 | 31.90 | 39.74 |   88.43 |  25.04 |  27.71 |  26.14 | 2 |true |
-| 47 | 25.57 | 31.45 | 34.41 |   89.86 |  25.57 |  25.51 |  26.12 | 2 |false |
-| 18 | 44.75 | 44.35 | 29.92 |   90.13 |  25.46 |  25.75 |  25.55 | 1 |false |
-| 63 | 43.95 | 25.03 | 27.92 |   90.35 |  29.94 |  25.03 |  25.05 | 1 |false |
-|  9 | 36.90 | 44.22 | 42.32 |   93.64 |  25.18 |  26.98 |  25.89 | 3 |false |
-|  1 | 27.71 | 37.43 | 35.56 |   94.02 |  26.65 |  28.64 |  25.21 | 0 |false |
-|  5 | 42.31 | 30.82 | 38.94 |   94.18 |  26.89 |  27.81 |  26.19 | 1 |false |
-|  6 | 40.96 | 44.17 | 39.78 |   94.81 |  33.00 |  27.19 |  30.08 | 0 |false |
-|  4 | 36.63 | 28.22 | 32.89 |   97.73 |  36.06 |  28.22 |  30.57 | 0 |true |
-| 55 | 34.76 | 38.24 | 41.16 |   98.47 |  26.18 |  27.42 |  25.58 | 1 |false |
-| 11 | 34.20 | 30.91 | 44.00 |   99.53 |  26.30 |  30.91 |  28.73 | 4 |false |
-|  3 | 25.01 | 44.13 | 37.25 |  101.67 |  25.01 |  26.20 |  26.09 | 2 |false |
