+++ Formula_one-main/main.c
@@ -8,7 +8,7 @@
 //TODO: add error handling/generating
 
 int checkArguments(int argc, char *argv[]) {
-    if (argc != 2 || (strcmp(argv[1], "-P1") != 0 && strcmp(argv[1], "-P2") != 0 && strcmp(argv[1], "-P3") != 0
+    if (argc != 2 || (strcmp(argv[1], "-h") != 0 && strcmp(argv[1], "-P1") != 0 && strcmp(argv[1], "-P2") != 0 && strcmp(argv[1], "-P3") != 0
             && strcmp(argv[1], "-Q1") != 0 && strcmp(argv[1], "-Q2") != 0 && strcmp(argv[1], "-Q3") != 0
             && strcmp(argv[1], "-sprint") != 0 && strcmp(argv[1], "-race") != 0)) {
         fprintf(stderr, "Error: %s [-P1 | -P2 | -P3 | -Q1 | -Q2 | -Q3 | -sprint | -race]\n", argv[0]);
@@ -28,6 +28,9 @@ int main(int argc, char *argv[]) {
         printf("  -P1     start the Practise 1 session (1H)\n");
         printf("  -P2     start the Practise 2 session (1H)\n");
         printf("  -P3     start the Practise 3 session (1H)\n");
+        printf("  -Q1     start the Qualifiying 1 session (12min)\n");
+        printf("  -Q2     start the Qualifiying 2 session (10min)\n");
+        printf("  -Q3     start the Qualifiying 3 session (8min)\n");
         printf("  -sprint     start the sprint race (100-120 km)\n");
         printf("  -race     start the sunday race (300-3500 km)\n");
         return 0;
@@ -49,28 +52,38 @@ int main(int argc, char *argv[]) {
     int num_laps;
     char* filename;
     if (strcmp(argv[1], "-P1") == 0) {
-        num_laps = 30;
+        num_laps = 35;
         filename = "P1.csv";
     } else if (strcmp(argv[1], "-P2") == 0) {
-        num_laps = 30;
+        num_laps = 35;
         filename = "P2.csv";
     } else if (strcmp(argv[1], "-P3") == 0) {
-        num_laps = 30;
+        num_laps = 35;
         filename = "P3.csv";
     } else if (strcmp(argv[1], "-Q1") == 0) {
-        num_laps = 12;
+        num_laps = 7;
         filename = "Q1.csv";
     } else if (strcmp(argv[1], "-Q2") == 0) {
-        num_laps = 8;
-        filename = "Q2.csv";
+        if (access("results/Q1.csv", F_OK) != -1) {
+            num_laps = 6;
+            filename = "Q2.csv";
+        } else {
+            fprintf(stderr, "Error: Q1 must be completed before Q2\n");
+            exit(EXIT_FAILURE);
+        }
     } else if (strcmp(argv[1], "-Q3") == 0) {
-        num_laps = 4;
-        filename = "Q3.csv";
+        if (access("results/Q2.csv", F_OK) != -1) {
+            num_laps = 5;
+            filename = "Q3.csv";
+        } else {
+            fprintf(stderr, "Error: Q2 must be completed before Q3\n");
+            exit(EXIT_FAILURE);
+        }
     } else if (strcmp(argv[1], "-sprint") == 0) {
-        num_laps = 20;
+        num_laps = 17;
         filename = "Sprint.csv";
     } else if (strcmp(argv[1], "-race") == 0) {
-        num_laps = 40;
+        num_laps = 52;
         filename = "Race.csv";
     }
 
@@ -89,8 +102,8 @@ int main(int argc, char *argv[]) {
         exit(EXIT_FAILURE);
     //printf("%d\n",num_cars);
 
-    static int carIds[] = {44, 63, 1, 11, 55, 16, 4, 3, 14, 31, 10, 22,
-                           5, 18, 6, 23, 77, 24, 47, 9};
+    static int carIds[] = {1, 11, 16, 55, 63, 44, 31, 10, 4, 81, 77, 24,
+                            14, 18, 20, 27, 22, 3, 23, 2};
     for (int i = 0; i < num_cars; i++) {
         init_car(&circuit[i], carIds[i]);
     }
@@ -158,9 +171,6 @@ int main(int argc, char *argv[]) {
         int num_qualified_cars = (strcmp(argv[1], "-Q2") == 0) ? 15 : 10;
         int* qualified_car_ids = read_qualified_car_ids_from_csv(filename, num_qualified_cars);
         init_circuit_with_qualified_car_ids(circuit, qualified_car_ids, num_qualified_cars);
-        for (int i = 0; i < num_qualified_cars; i++) {
-            printf("Qualified car ID %d: %d\n", i, qualified_car_ids[i]);
-        }
         free(qualified_car_ids);
     }
