+++ Formula_one-main/results/P1.csv
@@ -1,0 +1,24 @@
+P1.csv
+id; best s1; best s2 ; best s3 ; best lap; total time
+| id |   s1  |   s2  |   s3  | best lap | best s1 | best s2 | best s3 | pitstop | crash 
+|----|-------|-------|-------|----------|---------|---------|---------|---------|-------
+| 23 | 31.90 | 30.39 | 32.07 |   79.12 |  25.23 |  25.03 |  25.25 |11 |false |
+| 20 | 34.26 | 26.11 | 28.03 |   80.01 |  25.47 |  25.34 |  25.36 | 6 |true |
+|  3 | 42.91 | 42.30 | 30.47 |   80.27 |  25.27 |  25.31 |  25.25 | 3 |true |
+|  4 | 36.00 | 38.57 | 34.83 |   81.76 |  25.32 |  25.00 |  25.88 | 6 |true |
+| 81 | 36.09 | 34.89 | 32.57 |   83.00 |  25.01 |  25.24 |  25.44 | 4 |false |
+|  1 | 28.19 | 39.71 | 35.51 |   83.05 |  25.11 |  26.02 |  25.76 | 2 |true |
+| 55 | 34.00 | 44.84 | 38.51 |   83.35 |  25.51 |  25.67 |  25.17 | 5 |true |
+| 11 | 37.20 | 28.90 | 29.74 |   84.13 |  25.42 |  25.41 |  26.12 | 7 |false |
+| 14 | 35.44 | 31.01 | 37.59 |   86.24 |  25.36 |  29.34 |  25.71 | 0 |true |
+| 16 | 39.04 | 34.42 | 26.52 |   86.39 |  25.21 |  25.19 |  25.84 | 3 |false |
+| 10 | 39.74 | 35.20 | 29.55 |   86.58 |  25.25 |  25.10 |  25.64 | 0 |true |
+| 44 | 43.93 | 36.54 | 27.67 |   86.90 |  25.14 |  25.07 |  25.66 | 5 |true |
+| 77 | 41.12 | 41.24 | 27.59 |   86.97 |  25.09 |  25.60 |  25.24 | 5 |true |
+| 18 | 30.47 | 30.76 | 40.45 |   88.70 |  25.01 |  25.93 |  25.20 | 4 |false |
+|  2 | 41.81 | 42.47 | 37.38 |   89.95 |  26.63 |  25.38 |  25.13 | 1 |true |
+| 22 | 35.82 | 37.23 | 43.86 |   90.71 |  25.09 |  25.10 |  28.85 | 0 |true |
+| 63 | 35.91 | 40.97 | 28.84 |   92.86 |  25.39 |  27.34 |  26.04 | 2 |true |
+| 24 | 35.92 | 38.24 | 31.10 |   96.19 |  25.68 |  26.03 |  31.10 | 0 |true |
+| 31 | 27.59 | 39.57 | 35.99 |  103.14 |  27.59 |  39.57 |  35.99 | 0 |true |
+| 27 | 41.51 | 33.53 | 32.91 |  107.95 |  37.93 |  33.53 |  32.91 | 1 |true |
