+++ Formula_one-main/results/P2.csv
@@ -1,0 +1,24 @@
+P2.csv
+id; best s1; best s2 ; best s3 ; best lap; total time
+| id |   s1  |   s2  |   s3  | best lap | best s1 | best s2 | best s3 | pitstop | crash 
+|----|-------|-------|-------|----------|---------|---------|---------|---------|-------
+| 44 | 31.51 | 44.22 | 41.92 |   78.15 |  25.20 |  25.04 |  25.14 | 3 |false |
+| 31 | 29.55 | 30.49 | 35.33 |   79.51 |  25.19 |  25.13 |  25.09 | 8 |false |
+|  3 | 30.45 | 41.30 | 32.34 |   79.84 |  25.12 |  25.29 |  25.05 | 8 |false |
+| 55 | 38.72 | 28.84 | 44.92 |   81.25 |  26.45 |  25.06 |  25.97 | 2 |true |
+| 81 | 44.48 | 31.10 | 32.14 |   81.26 |  25.58 |  25.42 |  25.49 | 4 |true |
+| 11 | 35.55 | 33.12 | 32.20 |   82.92 |  25.15 |  25.93 |  25.22 | 1 |true |
+| 10 | 41.53 | 44.50 | 32.39 |   84.54 |  25.00 |  25.12 |  25.51 | 7 |false |
+| 16 | 40.98 | 42.44 | 44.33 |   84.89 |  25.10 |  25.02 |  25.13 | 7 |false |
+| 24 | 41.36 | 27.42 | 43.97 |   84.96 |  25.39 |  25.02 |  25.20 | 5 |false |
+| 22 | 35.84 | 43.51 | 26.30 |   85.00 |  25.02 |  25.09 |  25.44 | 5 |false |
+| 63 | 26.55 | 30.18 | 29.57 |   85.02 |  25.39 |  25.15 |  25.74 |11 |false |
+| 14 | 34.33 | 40.19 | 34.99 |   85.60 |  25.55 |  25.08 |  25.12 | 3 |false |
+| 20 | 34.78 | 30.28 | 40.00 |   85.90 |  25.04 |  26.52 |  25.47 | 2 |true |
+| 27 | 42.24 | 44.39 | 40.74 |   86.57 |  25.05 |  25.44 |  25.04 | 4 |false |
+| 77 | 40.34 | 31.62 | 33.27 |   86.98 |  25.00 |  25.08 |  25.89 | 6 |false |
+| 18 | 36.48 | 33.33 | 25.37 |   87.39 |  25.12 |  25.46 |  25.24 | 4 |false |
+|  1 | 33.49 | 27.88 | 26.81 |   88.18 |  25.56 |  26.80 |  26.78 | 1 |true |
+|  2 | 39.60 | 28.98 | 36.18 |   88.47 |  25.48 |  25.02 |  25.09 | 3 |false |
+| 23 | 27.61 | 39.71 | 42.44 |   90.47 |  25.40 |  25.90 |  25.61 | 4 |true |
+|  4 | 35.91 | 40.97 | 28.84 |   92.86 |  25.39 |  27.34 |  26.04 | 2 |true |
