+++ Formula_one-main/results/P3.csv
@@ -1,0 +1,24 @@
+P3.csv
+id; best s1; best s2 ; best s3 ; best lap; total time
+| id |   s1  |   s2  |   s3  | best lap | best s1 | best s2 | best s3 | pitstop | crash 
+|----|-------|-------|-------|----------|---------|---------|---------|---------|-------
+| 27 | 37.27 | 43.48 | 37.80 |   77.57 |  26.01 |  25.38 |  25.10 | 4 |false |
+|  2 | 42.70 | 34.17 | 40.83 |   78.02 |  25.59 |  25.66 |  25.31 | 1 |true |
+| 10 | 26.97 | 32.45 | 27.90 |   79.51 |  25.86 |  25.13 |  26.05 | 2 |true |
+|  3 | 29.34 | 29.72 | 38.24 |   79.71 |  25.11 |  25.93 |  25.34 | 3 |false |
+| 18 | 27.69 | 26.46 | 33.91 |   79.92 |  25.45 |  26.46 |  25.17 | 4 |true |
+| 24 | 34.97 | 42.94 | 38.61 |   80.32 |  25.40 |  25.22 |  25.30 | 1 |false |
+| 22 | 43.94 | 36.46 | 29.65 |   81.62 |  25.01 |  26.45 |  25.41 | 8 |false |
+| 63 | 35.12 | 29.80 | 33.84 |   81.81 |  25.22 |  26.02 |  26.00 | 0 |true |
+| 81 | 35.72 | 26.25 | 30.96 |   83.49 |  25.17 |  25.12 |  25.11 | 4 |false |
+| 23 | 43.48 | 28.59 | 31.74 |   83.79 |  25.36 |  25.36 |  25.09 | 8 |false |
+| 20 | 26.67 | 42.47 | 44.51 |   84.03 |  25.63 |  25.91 |  25.95 | 9 |true |
+|  1 | 28.29 | 28.73 | 33.27 |   84.13 |  25.74 |  25.16 |  25.44 | 4 |true |
+|  4 | 41.23 | 40.34 | 34.11 |   84.13 |  26.55 |  25.25 |  25.35 | 3 |true |
+| 44 | 37.34 | 26.26 | 28.64 |   84.54 |  25.71 |  25.23 |  25.65 | 5 |true |
+| 77 | 31.89 | 35.25 | 39.98 |   84.56 |  26.16 |  25.08 |  25.18 | 1 |true |
+| 11 | 39.36 | 31.34 | 29.24 |   86.28 |  25.29 |  25.07 |  25.53 | 6 |false |
+| 55 | 26.10 | 27.35 | 33.84 |   86.86 |  25.19 |  25.11 |  25.20 | 7 |false |
+| 31 | 25.39 | 29.94 | 36.94 |   88.52 |  25.39 |  25.15 |  26.12 | 2 |true |
+| 14 | 43.14 | 40.37 | 28.12 |   94.21 |  25.77 |  26.01 |  25.94 | 0 |true |
+| 16 | 40.88 | 44.01 | 31.99 |   98.98 |  26.02 |  30.05 |  30.81 | 0 |true |
