+++ Formula_one-main/results/Q1.csv
@@ -1,0 +1,24 @@
+Q1.csv
+id; best s1; best s2 ; best s3 ; best lap; total time
+| id |   s1  |   s2  |   s3  | best lap | best s1 | best s2 | best s3 | pitstop | crash 
+|----|-------|-------|-------|----------|---------|---------|---------|---------|-------
+| 27 | 35.49 | 27.88 | 40.25 |   81.18 |  27.01 |  25.32 |  26.36 | 1 |false |
+|  2 | 30.45 | 41.48 | 27.82 |   84.90 |  25.20 |  25.97 |  27.82 | 2 |false |
+| 63 | 36.05 | 33.43 | 35.25 |   86.26 |  27.27 |  25.78 |  26.96 | 1 |false |
+| 10 | 43.18 | 26.80 | 42.42 |   87.65 |  26.00 |  25.34 |  31.03 | 0 |true |
+| 31 | 36.05 | 31.14 | 33.93 |   87.65 |  30.46 |  29.04 |  25.01 | 2 |false |
+| 11 | 25.98 | 38.58 | 27.76 |   89.32 |  25.41 |  26.73 |  25.40 | 2 |false |
+| 14 | 32.06 | 35.68 | 44.54 |   89.83 |  31.64 |  25.37 |  27.14 | 1 |false |
+| 16 | 25.47 | 25.33 | 40.14 |   90.30 |  25.47 |  25.33 |  27.91 | 2 |false |
+|  3 | 42.70 | 26.27 | 29.75 |   91.14 |  26.64 |  26.27 |  26.17 | 2 |false |
+| 81 | 40.70 | 30.53 | 30.45 |   91.36 |  26.42 |  29.01 |  25.81 | 2 |false |
+| 20 | 26.76 | 43.48 | 42.52 |   91.50 |  26.76 |  25.69 |  30.23 | 0 |false |
+| 44 | 34.66 | 36.83 | 34.79 |   91.52 |  26.61 |  29.06 |  25.19 | 1 |false |
+|  1 | 26.80 | 37.23 | 28.20 |   92.24 |  26.74 |  25.19 |  27.18 | 2 |true |
+| 77 | 34.07 | 31.18 | 27.00 |   92.25 |  27.24 |  28.56 |  27.00 | 0 |false |
+| 55 | 39.85 | 41.22 | 32.24 |   93.79 |  26.43 |  26.82 |  26.80 | 1 |false |
+| 18 | 38.25 | 43.53 | 26.25 |   93.81 |  26.40 |  28.43 |  26.25 | 0 |true |
+| 24 | 29.92 | 33.52 | 34.02 |   96.78 |  28.68 |  26.83 |  29.91 | 1 |false |
+| 23 | 31.27 | 38.88 | 27.29 |   97.44 |  31.27 |  38.88 |  27.29 | 0 |true |
+|  4 | 29.34 | 37.44 | 38.98 |   98.18 |  29.34 |  31.14 |  26.68 | 1 |true |
+| 22 | 33.85 | 44.79 | 32.52 |  111.16 |  33.85 |  44.79 |  32.52 | 0 |true |
