+++ Formula_one-main/results/Q2.csv
@@ -1,0 +1,19 @@
+Q2.csv
+id; best s1; best s2 ; best s3 ; best lap; total time
+| id |   s1  |   s2  |   s3  | best lap | best s1 | best s2 | best s3 | pitstop | crash 
+|----|-------|-------|-------|----------|---------|---------|---------|---------|-------
+| -1094795586 | 26.26 | 30.01 | 26.30 |   82.57 |  26.26 |  27.39 |  25.24 | 0 |false |
+| -1094795586 | 38.48 | 35.52 | 42.71 |   88.02 |  29.93 |  27.54 |  25.94 | 2 |false |
+| -1094795586 | 37.88 | 29.68 | 34.18 |   89.55 |  26.26 |  25.75 |  34.18 | 0 |false |
+| -1094795586 | 33.82 | 30.20 | 34.02 |   90.47 |  28.49 |  28.50 |  28.42 | 0 |false |
+| -1094795586 | 42.43 | 41.57 | 38.79 |   90.99 |  29.43 |  25.79 |  33.76 | 1 |true |
+| -1094795586 | 37.08 | 34.02 | 30.79 |   92.63 |  26.10 |  27.86 |  26.72 | 0 |false |
+| -1094795586 | 39.07 | 44.16 | 39.52 |   92.91 |  28.40 |  30.90 |  32.64 | 2 |false |
+| -1094795586 | 35.00 | 32.81 | 25.20 |   93.02 |  31.82 |  25.69 |  25.20 | 1 |false |
+| -1094795586 | 36.98 | 26.19 | 39.83 |   95.19 |  25.34 |  26.19 |  26.44 | 2 |false |
+| -1094795586 | 32.41 | 30.84 | 39.40 |   96.27 |  26.79 |  25.76 |  30.55 | 1 |false |
+| -1094795586 | 32.46 | 44.98 | 37.18 |  101.03 |  25.51 |  37.94 |  26.04 | 0 |false |
+| -1094795586 | 29.16 | 31.16 | 43.53 |  103.86 |  29.16 |  30.17 |  36.27 | 2 |true |
+| -1094795586 | 38.22 | 27.24 | 39.23 |  104.69 |  26.17 |  27.24 |  37.29 | 1 |false |
+| -1094795586 | 41.62 | 31.28 | 34.24 |  104.70 |  28.58 |  31.28 |  26.42 | 0 |false |
+| -1094795586 | 37.14 | 40.40 | 39.65 |  109.11 |  35.94 |  34.25 |  38.93 | 0 |true |
