+++ Formula_one-main/results/Q3.csv
@@ -1,0 +1,14 @@
+Q3.csv
+id; best s1; best s2 ; best s3 ; best lap; total time
+| id |   s1  |   s2  |   s3  | best lap | best s1 | best s2 | best s3 | pitstop | crash 
+|----|-------|-------|-------|----------|---------|---------|---------|---------|-------
+| -1094795586 | 39.72 | 31.77 | 32.25 |   81.63 |  25.27 |  25.17 |  25.42 | 0 |false |
+|  9 | 28.94 | 34.74 | 25.62 |   89.30 |  28.64 |  30.72 |  25.62 | 2 |false |
+| -1094795586 | 25.57 | 35.82 | 28.00 |   89.39 |  25.57 |  25.85 |  28.00 | 0 |false |
+| -1094795586 | 29.85 | 26.47 | 34.46 |   90.79 |  27.63 |  26.47 |  31.30 | 0 |false |
+|  6 | 36.29 | 43.36 | 30.00 |   91.48 |  25.12 |  29.68 |  30.00 | 0 |true |
+|  8 | 25.87 | 29.96 | 44.47 |   92.50 |  25.87 |  27.91 |  26.85 | 0 |false |
+|  7 | 36.74 | 31.56 | 25.50 |   93.80 |  29.58 |  27.20 |  25.50 | 1 |false |
+| -1094795586 | 33.19 | 30.58 | 33.35 |   97.12 |  33.19 |  25.34 |  33.35 | 2 |false |
+| 10 | 34.74 | 33.20 | 42.18 |   98.35 |  34.74 |  25.28 |  28.24 | 0 |false |
+|  5 | 30.86 | 34.03 | 35.04 |   99.52 |  30.86 |  27.71 |  28.95 | 1 |false |
