+++ Formula_one-main/results/Race.csv
@@ -1,0 +1,24 @@
+Race.csv
+id; best s1; best s2 ; best s3 ; best lap; total time
+| id |   s1  |   s2  |   s3  | best lap | best s1 | best s2 | best s3 | pitstop | crash 
+|----|-------|-------|-------|----------|---------|---------|---------|---------|-------
+| 22 | 30.03 | 41.01 | 25.55 |   78.25 |  25.28 |  25.06 |  25.16 | 8 |false |
+| 14 | 34.67 | 29.53 | 37.50 |   79.85 |  25.08 |  25.33 |  25.15 | 5 |true |
+| 55 | 28.36 | 33.08 | 42.11 |   80.63 |  25.37 |  25.14 |  25.39 |10 |true |
+| 63 | 43.29 | 43.76 | 39.82 |   81.01 |  25.26 |  25.02 |  25.03 | 8 |true |
+| 10 | 30.33 | 25.47 | 25.31 |   81.12 |  25.82 |  25.39 |  25.31 | 0 |true |
+| 24 | 29.72 | 32.58 | 34.63 |   81.33 |  25.21 |  25.06 |  25.41 | 9 |false |
+| 20 | 40.32 | 35.63 | 25.66 |   81.62 |  25.20 |  25.21 |  25.65 | 3 |true |
+|  2 | 42.36 | 38.08 | 32.41 |   81.76 |  25.39 |  25.06 |  25.53 | 5 |true |
+|  4 | 39.81 | 41.07 | 38.02 |   82.20 |  25.26 |  25.03 |  25.19 | 5 |true |
+| 31 | 30.27 | 44.69 | 28.10 |   82.35 |  25.27 |  25.09 |  25.40 |11 |false |
+| 44 | 26.23 | 41.68 | 34.90 |   83.49 |  26.23 |  25.09 |  25.11 | 3 |true |
+| 18 | 34.51 | 43.05 | 29.37 |   84.00 |  25.19 |  25.11 |  25.32 | 3 |true |
+| 23 | 31.87 | 27.88 | 25.45 |   84.31 |  25.45 |  25.08 |  25.45 |12 |true |
+|  3 | 30.13 | 25.09 | 29.59 |   84.80 |  25.56 |  25.09 |  25.06 | 6 |true |
+|  1 | 28.90 | 28.80 | 32.18 |   86.56 |  25.11 |  25.55 |  25.09 | 1 |true |
+| 77 | 40.44 | 39.55 | 40.77 |   86.90 |  26.41 |  25.46 |  26.33 | 1 |true |
+| 27 | 30.02 | 32.78 | 35.43 |   87.24 |  26.87 |  31.94 |  27.61 | 1 |true |
+| 16 | 29.61 | 44.75 | 43.73 |   90.85 |  25.22 |  30.07 |  30.83 | 2 |true |
+| 11 | 39.39 | 27.01 | 30.94 |   91.99 |  25.03 |  26.47 |  28.57 | 0 |true |
+| 81 | 38.65 | 36.77 | 33.14 |   96.00 |  25.72 |  27.33 |  27.49 | 0 |true |
