+++ Formula_one-main/results/Sprint.csv
@@ -1,0 +1,24 @@
+Sprint.csv
+id; best s1; best s2 ; best s3 ; best lap; total time
+| id |   s1  |   s2  |   s3  | best lap | best s1 | best s2 | best s3 | pitstop | crash 
+|----|-------|-------|-------|----------|---------|---------|---------|---------|-------
+| 16 | 39.81 | 30.06 | 30.36 |   80.81 |  25.11 |  25.55 |  25.32 | 1 |false |
+|  2 | 42.05 | 43.75 | 31.83 |   83.79 |  25.56 |  25.10 |  25.36 | 3 |false |
+| 22 | 26.42 | 44.49 | 39.00 |   84.56 |  26.42 |  25.31 |  25.65 | 5 |false |
+| 55 | 31.89 | 35.25 | 39.98 |   84.56 |  26.16 |  25.08 |  25.18 | 1 |true |
+| 11 | 28.82 | 30.59 | 25.34 |   84.75 |  28.12 |  25.61 |  25.34 | 4 |true |
+|  4 | 39.17 | 32.01 | 25.98 |   84.77 |  25.95 |  27.58 |  25.98 | 1 |true |
+| 27 | 43.53 | 37.51 | 34.13 |   85.62 |  26.63 |  25.15 |  25.03 | 1 |true |
+| 23 | 30.53 | 26.73 | 36.25 |   86.04 |  25.63 |  26.73 |  25.95 | 3 |false |
+| 81 | 37.64 | 44.11 | 41.69 |   86.62 |  26.90 |  25.70 |  25.08 | 0 |true |
+|  1 | 26.44 | 28.63 | 43.11 |   87.49 |  25.46 |  26.30 |  25.41 | 3 |false |
+| 20 | 27.75 | 42.13 | 28.74 |   88.96 |  26.08 |  26.43 |  28.45 | 3 |true |
+| 44 | 37.97 | 42.65 | 26.86 |   89.01 |  25.10 |  26.85 |  26.86 | 3 |true |
+| 18 | 33.95 | 32.52 | 31.70 |   89.17 |  26.39 |  25.25 |  25.02 | 3 |true |
+| 24 | 32.95 | 29.87 | 26.36 |   89.19 |  26.27 |  26.26 |  26.36 | 3 |false |
+|  3 | 41.11 | 39.83 | 27.16 |   90.12 |  26.24 |  25.78 |  25.37 | 4 |false |
+| 31 | 29.61 | 44.75 | 43.73 |   90.85 |  25.22 |  30.07 |  30.83 | 2 |true |
+| 10 | 35.49 | 36.78 | 33.97 |   91.41 |  25.05 |  27.15 |  25.32 | 1 |false |
+| 63 | 29.60 | 35.41 | 35.59 |   91.75 |  25.53 |  26.46 |  26.25 | 4 |true |
+| 77 | 33.76 | 30.51 | 43.99 |   92.10 |  25.17 |  25.15 |  38.86 | 0 |true |
+| 14 | 27.95 | 37.37 | 31.34 |   96.67 |  27.95 |  37.37 |  31.34 | 0 |true |
